requirements.txt
Flask>=3.0,<4.0
python-dotenv>=1.0,<2.0

pip~=26.0.1
zipp~=3.23.0
markupsafe~=3.0.3
werkzeug~=3.1.6
click~=8.1.8
jinja2~=3.1.6
blinker~=1.9.0
itsdangerous~=2.2.0
wheel~=0.46.2
setuptools~=78.1.1